requirements.txt
asgiref==3.5.2
boto3==1.26.31
botocore==1.29.31
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
dj-rest-auth==2.2.5
Django==4.1.2
django-allauth==0.51.0
django-rest-auth==0.9.5
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
future==0.18.2
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.5
requests==2.28.1
requests-oauthlib==1.3.1
s3transfer==0.6.0
serializers==0.2.4
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
